mums/requirements/common.txt
Django==1.9.9
djangorestframework==3.4.6
